+++ lua/luau-lsp/context.lua
@@ -1,10 +1,6 @@
-local log = require "luau-lsp.log"
-local util = require "luau-lsp.util"
-
 local FFLAG_KINDS = { "FFlag", "FInt", "DFFlag", "DFInt" }
 
 ---@class luau-lsp.Context
----@field cmd string
 ---@field fflags table<string, string>
 ---@field definitions table<string, string>
 ---@field documentation table<integer, string>
@@ -40,25 +36,12 @@ function Context:add_definitions(name, path)
   if not name:find "^@" then
     name = "@" .. name
   end
-
-  if util.is_file(path) then
-    self.definitions[name] = path
-  else
-    log.warn(
-      "Definitions file '%s' at '%s' does not exist, types will not be provided from this file",
-      name,
-      path
-    )
-  end
+  self.definitions[name] = path
 end
 
 ---@param path string
 function Context:add_documentation(path)
-  if util.is_file(path) then
-    table.insert(self.documentation, path)
-  else
-    log.warn("Documentation file at '%s' does not exist", path)
-  end
+  table.insert(self.documentation, path)
 end
 
 return Context